apps/doc-service/pyproject.toml
[project]
name = "fluidfill-doc-service"
version = "0.1.0"
description = "FastAPI service for parsing/rendering legal docs"
requires-python = ">=3.11"
dependencies = [
  "fastapi==0.111.0",
  "uvicorn[standard]==0.30.0",
  "python-docx==1.1.2",
  "pydantic==2.7.1"
]
